Lab2/src/main/java/Labs_OOP_sem_3/operations/LeftSteppingDifferentialOperator.java
package Labs_OOP_sem_3.operations;

import Labs_OOP_sem_3.functions.MathFunction;

public class LeftSteppingDifferentialOperator extends SteppingDifferentialOperator {

    public LeftSteppingDifferentialOperator(double step) {
        super(step);
    }

    @Override
    public MathFunction derive(MathFunction function) {
        return new MathFunction() {
            @Override
            public double apply(double x) {
                return (function.apply(x) - function.apply(x - step)) / step;
            }
        };
    }

    @Override
    public double apply(double x) {
        return 0;
    }
}
